feat: created enum dictionaries and types for the new data for better typing and autocomplete assistance
