+++ crypto.sap.ui5/webapp/types/Dictionaries.ts
@@ -1,0 +1,25 @@
+import TopMarketCap from "../controller/TopMarketCap.controller";
+
+export enum ModelNamesDictionary {
+  cryptoModel = "cryptoModel",
+  topMarketCap = "topMarketCap",
+  cryptoDetail = "cryptoDetail"
+}
+
+// Crypto Model Properties
+export enum CryptoModelPropsDictionary {
+  supportedCurrencies = "/supportedCurrencies",
+  selectedCurrency = "/selectedCurrency",
+}
+
+// Top Market Cap Model properties
+export enum TMCModelPropsDictionary {
+  page = "/page",
+  crypto = "/crypto",
+}
+
+export enum CryptoDetailModelPropsDictionary {
+  cryptoDetail = '/cryptoDetail',
+  selectedDays = "/selectedDays",
+  priceAxisRange = "/priceAxisRange",
+}
